Rename DeepScaner to DeepScanner
