+++ credsweeper/deep_scanner/deep_scanner.py
